refactor: family id Integer 타입으로 변경

+++ backend/src/main/java/org/ssafy/eeum/domain/family/dto/FamilyDetailResponseDto.java
@@ -11,7 +11,7 @@ import java.util.List;
 @Schema(description = "가족 상세 조회 응답 DTO")
 public class FamilyDetailResponseDto {
     @Schema(description = "가족 ID")
-    private Long familyId;
+    private Integer familyId;
 
     @Schema(description = "가족 이름")
     private String groupName;
